Moviendo las promesas a location.ts
