+++ frontend/www/js/omegaup/arena/location.ts
@@ -1,6 +1,9 @@
 import { types } from '../api_types';
 import { PopupDisplayed } from '../components/problem/Details.vue';
 import clarificationsStore from './clarificationsStore';
+import * as api from '../api';
+import { trackRun } from './submissions';
+import problemsStore from './problemStore';
 
 export interface LocationOptions {
   problem: types.NavbarProblemsetProblem | null;
@@ -62,4 +65,51 @@ export function getOptionsFromLocation(location: string): LocationOptions {
   }
 
   return response;
+}
+
+export async function getProblemAndRunDetails({
+  location,
+  problems,
+}: {
+  location: string;
+  problems?: types.NavbarProblemsetProblem[];
+}): Promise<{
+  runDetails: null | types.RunDetails;
+  problemDetails: null | types.ProblemDetails;
+}> {
+  const { guid, problemAlias } = getOptionsFromLocation(location);
+  let runDetails: null | types.RunDetails = null;
+  let problemDetails: null | types.ProblemDetails = null;
+  let problemPromise: Promise<null | types.ProblemDetails> = Promise.resolve(
+    null,
+  );
+  let runPromise: Promise<null | types.RunDetails> = Promise.resolve(null);
+
+  if (problemAlias) {
+    problemPromise = api.Problem.details({
+      problem_alias: problemAlias,
+      prevent_problemset_open: false,
+    });
+  }
+  if (guid) {
+    runPromise = api.Run.details({ run_alias: guid });
+  }
+
+  [problemDetails, runDetails] = await Promise.all([
+    problemPromise,
+    runPromise,
+  ]);
+
+  if (problemDetails != null) {
+    for (const run of problemDetails.runs ?? []) {
+      trackRun({ run });
+    }
+    const currentProblem = problems?.find(
+      ({ alias }: { alias: string }) => alias === problemDetails?.alias,
+    );
+    problemDetails.title = currentProblem?.text ?? '';
+    problemsStore.commit('addProblem', problemDetails);
+  }
+
+  return { problemDetails, runDetails };
 }
